+++ packages/react-native-the-sheet/src/core/bottom-sheet/bottom-sheet-provider.tsx
@@ -1,5 +1,5 @@
 import { createContext, useContext, useEffect, useId, useRef } from 'react'
-import Animated, {
+import {
   useSharedValue,
   useDerivedValue,
   useAnimatedRef,
@@ -7,7 +7,11 @@ import Animated, {
 import { useSyncedSharedValue } from '../hooks/use-synced-shared-value'
 import { useToSharedValue } from '../hooks/use-to-shared-value'
 import { useTrueSafeArea } from '../hooks'
-import type { BottomSheetContextType, BottomSheetProviderProps } from './types'
+import type {
+  BottomSheetContextType,
+  BottomSheetProviderProps,
+  ScrollViewRefCore,
+} from './types'
 import { useBottomSheetRegistry } from './bottom-sheet-registry-provider'
 
 const BottomSheetContext = createContext<BottomSheetContextType | undefined>(
@@ -89,9 +93,7 @@ export function BottomSheetProvider({
 
   // MARK: Bottom sheet context
 
-  const scrollViewRef = useAnimatedRef<
-    Animated.ScrollView | Animated.FlatList<unknown>
-  >()
+  const scrollViewRef = useAnimatedRef<ScrollViewRefCore>()
   const isScrollViewReady = useSharedValue(false)
   const isScrollViewInteracting = useSharedValue(false)
   const scrollY = useSharedValue(0)