+++ packages/react-native-the-sheet/src/core/bottom-sheet/types.ts
@@ -1,18 +1,24 @@
-import type { ComponentProps, PropsWithChildren } from 'react'
+import type { ComponentProps, ComponentRef, PropsWithChildren } from 'react'
 import type {
+  DefaultSectionT,
   GestureResponderEvent,
   LayoutChangeEvent,
+  SectionListProps,
   StyleProp,
   ViewStyle,
+  VirtualizedListProps,
 } from 'react-native'
 import type Animated from 'react-native-reanimated'
 import type {
+  AnimatedProps,
   AnimatedRef,
   ScrollHandler,
   SharedValue,
 } from 'react-native-reanimated'
 import type { AnimatedProp } from '../types'
 import type { PanGesture } from 'react-native-gesture-handler'
+import type { AnimatedVirtualizedList } from './bottom-sheet-virtualized-list'
+import type { AnimatedSectionList } from './bottom-sheet-section-list'
 
 // MARK: Bottom sheet handle
 
@@ -42,6 +48,12 @@ export type BottomSheetRegistryProviderProps = PropsWithChildren
 /** Percentage is compared to the screen height */
 export type SnapPoint = number | `${number}%`
 
+export type ScrollViewRefCore =
+  | Animated.ScrollView
+  | Animated.FlatList<unknown>
+  | ComponentRef<typeof AnimatedVirtualizedList>
+  | ComponentRef<typeof AnimatedSectionList>
+
 export type BottomSheetContextType = {
   enableFloat: SharedValue<boolean>
   enableOverdrag: SharedValue<boolean>
@@ -56,7 +68,7 @@ export type BottomSheetContextType = {
   translateY: SharedValue<number>
   isTranslateYAnimating: SharedValue<boolean>
 
-  scrollViewRef: AnimatedRef<Animated.ScrollView | Animated.FlatList<unknown>>
+  scrollViewRef: AnimatedRef<ScrollViewRefCore>
   isScrollViewReady: SharedValue<boolean>
   isScrollViewInteracting: SharedValue<boolean>
   scrollY: SharedValue<number>
@@ -145,6 +157,74 @@ type AnimatedFlatListProps<T> = ComponentProps<typeof Animated.FlatList<T>>
 
 export type BottomSheetFlatListProps<T> = Omit<
   AnimatedFlatListProps<T>,
+  | 'onLayout'
+  | 'onContentSizeChange'
+  | 'onTouchStart'
+  | 'onTouchEnd'
+  | 'onScroll'
+  | 'onBeginDrag'
+  | 'onEndDrag'
+  | 'onMomentumBegin'
+  | 'onMomentumEnd'
+> & {
+  fill?: AnimatedProp<boolean>
+  getPanGesture?: () => PanGesture
+
+  onLayout?: (e: LayoutChangeEvent) => void
+  onContentSizeChange?: (w: number, h: number) => void
+  onTouchStart?: (e: GestureResponderEvent) => void
+  onTouchEnd?: (e: GestureResponderEvent) => void
+
+  onScroll?: ScrollHandler
+  onBeginDrag?: ScrollHandler
+  onEndDrag?: ScrollHandler
+  onMomentumBegin?: ScrollHandler
+  onMomentumEnd?: ScrollHandler
+}
+
+// MARK: Bottom sheet virtualized list
+
+type AnimatedVirtualizedListProps<T> = AnimatedProps<VirtualizedListProps<T>>
+
+export type BottomSheetVirtualizedListProps<T> = Omit<
+  AnimatedVirtualizedListProps<T>,
+  | 'onLayout'
+  | 'onContentSizeChange'
+  | 'onTouchStart'
+  | 'onTouchEnd'
+  | 'onScroll'
+  | 'onBeginDrag'
+  | 'onEndDrag'
+  | 'onMomentumBegin'
+  | 'onMomentumEnd'
+> & {
+  fill?: AnimatedProp<boolean>
+  getPanGesture?: () => PanGesture
+
+  onLayout?: (e: LayoutChangeEvent) => void
+  onContentSizeChange?: (w: number, h: number) => void
+  onTouchStart?: (e: GestureResponderEvent) => void
+  onTouchEnd?: (e: GestureResponderEvent) => void
+
+  onScroll?: ScrollHandler
+  onBeginDrag?: ScrollHandler
+  onEndDrag?: ScrollHandler
+  onMomentumBegin?: ScrollHandler
+  onMomentumEnd?: ScrollHandler
+}
+
+// MARK: Bottom sheet section list
+
+type AnimatedSectionListProps<
+  ItemT,
+  SectionT = DefaultSectionT,
+> = AnimatedProps<SectionListProps<ItemT, SectionT>>
+
+export type BottomSheetSectionListProps<
+  ItemT,
+  SectionT = DefaultSectionT,
+> = Omit<
+  AnimatedSectionListProps<ItemT, SectionT>,
   | 'onLayout'
   | 'onContentSizeChange'
   | 'onTouchStart'