+++ packages/react-native-the-sheet/CHANGELOG.md
@@ -1,5 +1,11 @@
 # react-native-the-sheet
 
+## 1.0.17
+
+### Patch Changes
+
+- Add `BottomSheetSectionList` and `BottomSheetVirtualizedList` components
+
 ## 1.0.16
 
 ### Patch Changes